+++ app/src/basic/res/layout/activity_main.xml
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/fragment2"
-        class="converter.Fragment_Buttons"
+        class="converter.Fragment_Buttons_basic"
         android:layout_width="match_parent"
         android:layout_height="692dp"
         android:layout_marginTop="0dp"
