+++ app/src/main/res/layout-land/activity_main.xml
@@ -24,7 +24,7 @@
 
     <fragment
         android:id="@+id/fragment2"
-        class="converter.Fragment_Buttons"
+        class="converter.Fragment_Buttons_basic"
         android:layout_width="390dp"
         android:layout_height="550dp"
         android:layout_below="@+id/TextView"
