+++ app/src/pro/java/converter/Fragment_Buttons.java
@@ -16,83 +16,18 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.converter.R;
 
-public class Fragment_Buttons extends Fragment  {
+public class Fragment_Buttons extends converter.Fragment_Buttons_basic {
     public ClipboardManager myClipboard;
-    SomeEventListener sel;
+
     private LogicVM myVM;
 
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        super.onCreateView(inflater,container,savedInstanceState);
         View v = inflater.inflate(R.layout.fragment_buttons,null);
         myClipboard = (ClipboardManager)getActivity().getSystemService(Context.CLIPBOARD_SERVICE);
         Button button;
-        button = (Button) v.findViewById(R.id.button1);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("1");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button2);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("2");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button3);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("3");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button4);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("4");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button5);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("5");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button6);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("6");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button7);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("7");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button8);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("8");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button9);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("9");
-            }
-        });
-        button = (Button) v.findViewById(R.id.button0);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("0");
-            }
-        });
-        button = (Button) v.findViewById(R.id.buttonBack);
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                myVM.setNumber("Back");
-            }
-        });
         button = (Button) v.findViewById(R.id.buttonCopy);
         button.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
@@ -108,8 +43,6 @@ public class Fragment_Buttons extends Fragment  {
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        myVM = ViewModelProviders.of(getActivity()).get(LogicVM.class);
-
     }
 
     @Override
@@ -132,17 +65,6 @@ public class Fragment_Buttons extends Fragment  {
         super.onStop();
     }
 
-    @Override
-    public void onAttach(Activity activity) {
-        super.onAttach(activity);
-        try{
-            sel = (SomeEventListener) activity;
-        }
-        catch (ClassCastException e)
-        {
-            throw new ClassCastException(activity.toString() + "must implement SomeEventListener");
-        }
-    }
 
 
 }
